backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-cloud-texttospeech>=2.16.3
pydantic>=2.7.0
python-dotenv>=1.0.0
